Override predict.py with content of fastmlapi_server.py
