+++ predict.py
@@ -1,93 +1,52 @@
-import time
-
 import torchvision
-
-import matplotlib.pyplot as plt
+from fastmlapi import MLController, postprocessing, prediction
 import numpy as np
+
+from Model import MyModel
 import torch
 
-import traceback
-import os
-from Model import MyModel
-
-PATH = "trained_models/model.pth"
-DEBUG = False
-
-tranform = torchvision.transforms.Compose([torchvision.transforms.ToTensor()])
-
-model = MyModel()
-
-model.load_state_dict(torch.load(PATH, weights_only=True))
+transform = torchvision.transforms.Compose([torchvision.transforms.ToTensor()])
 
 
-INPUT_PIPE_NAME = "model_input.pipe"
-OUTPUT_PIPE_NAME = "model_output.pipe"
+class ClassifierServer(MLController):
+    model_name = "handwritten-digits-recognizer"
+    model_version = "1.0.0"
 
-if not os.path.exists(INPUT_PIPE_NAME):
-    os.mkfifo(INPUT_PIPE_NAME)
+    def load_model(self):
+        PATH = "trained_models/model.pth"
+        model = MyModel()
+        model.load_state_dict(torch.load(PATH, weights_only=True))
+        return model
 
-if not os.path.exists(OUTPUT_PIPE_NAME):
-    os.mkfifo(OUTPUT_PIPE_NAME)
+    @prediction
+    def preprocess_and_predict(self, data: torch.Tensor) -> torch.Tensor:
+        X = np.array(data, dtype=np.float32) / 255.0
+        X = X.reshape((-1, 1, 28, 28))
 
-
-print("==================================")
-print("   Model loaded, listening...")
-print("=================================\n")
-
-
-while True:
-    with open(INPUT_PIPE_NAME, 'r') as input_file:
-        pred_start_time = time.time()
-
-        input_data = input_file.read().rstrip('\n')
-        input_data = input_data.split(',')
-
-        # print("Strings array: ", input_data)
-
-        try:
-            X = np.array(input_data, dtype=np.float32) / 255.0
-            X = X.reshape((28,28))
-        except ValueError:
-            # Wrong size or something...
-            traceback.print_exc()
-            continue
-        
-        # print(f"Received: {X.shape} {X}")
+        X_tensor = torch.from_numpy(X)
 
         with torch.no_grad():
-            x_tensor = torch.unsqueeze(tranform(X),0)
+            self.model.eval()
+            output = self.model(X_tensor)
+            return output
 
-            #print(f"{x_tensor=}")
-            output=model(x_tensor)
+    @postprocessing
+    def postprocess(self, predictions) -> list:
+        predictions: np.ndarray = np.squeeze(np.exp(predictions.numpy()))
+        if len(predictions.shape) == 1:
+            predictions = predictions.reshape(1, -1)
 
-            model.eval()
-            predictions: np.ndarray = np.squeeze(np.exp(output.numpy()))
-            
-        predicted_labels = np.argmax(predictions, axis=0)
+        predicted_labels = np.argmax(predictions, axis=1)
 
-        
-        
-        if DEBUG:
-            model_predictions = {
-            str(i): round(predictions[i].item(), 3)
-            for i in range(10)
+        response = [
+            {
+                "label": predicted_labels[i].tolist(),
+                "proba": predictions[i].tolist()
             }
-            print(f"\n\n{model_predictions=}")
+            for i in range(len(predicted_labels))
+        ]
 
-        pred_end_time = time.time()
-        
-        if DEBUG:
-            print(f"\nPrediction duration: {pred_end_time - pred_start_time} seconds")
+        return response
 
-        predictions: list[str] = (predictions * 255).flatten().astype(int).astype(str).tolist()
-        predictions = ",".join(predictions)
-
-        with open(OUTPUT_PIPE_NAME, "w") as output_file:
-            output_file.write(predictions)
-            output_file.flush()
-
-        if DEBUG:
-            plt.imshow(X.reshape(28, 28), cmap='gray')
-            plt.title(f"Predicted: {predicted_labels}")
-            plt.axis('off')
-            plt.savefig("debug_inputs/last.png")
+if __name__ == "__main__":
+    ClassifierServer().run()
